requirements.txt
python-decouple
elevenlabs